requirements.txt
git+https://github.com/stanislawbartkowski/ksef_pyth.git
pyyaml